requirements.txt
boto3==1.9.183
botocore==1.12.183
certifi==2019.6.16
chardet==3.0.4
docutils==0.14
ecdsa==0.13.2
enum-compat==0.0.2
future==0.17.1
idna==2.8
intuit-oauth==1.2.2
jmespath==0.9.4
oauthlib==3.0.2
pyasn1==0.4.5
pycparser==2.18
python-dateutil==2.8.0
python-jose==3.0.1
python-quickbooks==0.8.0
rauth==0.7.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
urllib3==1.25.3
